webinar_part_2/2.0_nxos_existing_with_netbox/requirements.txt
pynetbox
pyyaml
netaddr
ansible-vault
flask
pythondialog
napalm
